Update update_manager.py to hopefully add unavailable devices on failures
Previously it was sort of hiding errors.  But now it reports them.  When an update failed, the code would log a warning and then return the last known good data. The DataUpdateCoordinator saw a successful return and thought everything was fine. Now, a failure will result in devices being shown as unavailable.